Add test project AutoCompleteTest
My purpose is to debug the performance problem populating the dropdown list for the first time. This takes much longer the populating the list on subsequent occasions.

This test fills a list of words from the DataMuse web api, with a word count of 1000. With only 100, you don't notice the performance problem much, but with 1000 it is significant.