Minor changes to test program.
Minor changes to the test program. Set MAXRECORDS to 100 instead of 1000 and add a comment.
With MAXRECORDS = 1000, the initial display of the drop down list takes about 6 to 8 seconds.
Subsequent updates do not take so long. I cannot find a problem in the AutoCompleteBox and
I am now inclined to think that it is a WPF problem.