More updates to test program.
Add a second AutoCompleteBox using a simple list of string, and no DataTemplate.
This has exactly the same performance problem as the existing AutoCompleteBox.

Add a stupid logic to populate the list with 100 items for the first query, but
then 1000 items for all subsequent queries. This is a substantial improvement!