Another simple test program
This is another simple test program to demonstrate the performance problem in the AutoCompleteBox, opening the drop down list with 1000 items for the first time.
The test contains two identical controls, each with a list of 1000 items. The first one you open takes a long time. The second one is fast and all subsequent operations are fast. This illustrates that whatever is causing the performance problem is independent of the  specific control.